changed placement of tests
